Setting Up Logger
rather than using System.out.println(); for EVERYTHING I figured setting
up a logger to use the already built in FMLLog system would be much
cleaner. This is the ground work for a logger and I will be converting
all current debug commands to the logger in a later update.